requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.2
django-admin-list-charts==0.1.3
django-appconf==1.0.6
django-bootstrap5==23.4
django-imagekit==5.0.0
django-import-export==3.3.7
django-jazzmin==2.6.0
django-phonenumber-field==7.3.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
pilkit==3.0
pillow==10.2.0
psycopg2-binary==2.9.9
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.9.0
xlrd==2.0.1
xlwt==1.3.0
